+++ Python/Product/Analysis/Analysis.csproj
@@ -108,12 +108,13 @@
     <Compile Include="LanguageServer\DiagnosticsErrorSink.cs" />
     <Compile Include="LanguageServer\DocumentationBuilder.cs" />
     <Compile Include="LanguageServer\Enums.cs" />
-    <Compile Include="LanguageServer\Hooks.cs" />
+    <Compile Include="LanguageServer\Extensibility.cs" />
     <Compile Include="LanguageServer\ILanguageServerExtension.cs" />
     <Compile Include="LanguageServer\ILogger.cs" />
     <Compile Include="LanguageServer\MarkdownDocumentationBuilder.cs" />
     <Compile Include="LanguageServer\PlainTextDocumentationBuilder.cs" />
     <Compile Include="LanguageServer\Server.Completion.cs" />
+    <Compile Include="LanguageServer\Server.Extensions.cs" />
     <Compile Include="LanguageServer\Server.FindReferences.cs" />
     <Compile Include="LanguageServer\Server.Hover.cs" />
     <Compile Include="LanguageServer\Server.PrivateHelpers.cs" />
